Moved git related code to git.py
