+++ cfbs/commands.py
@@ -5,7 +5,6 @@ in main.py for -h/--help/help.
 import os
 import logging as log
 import json
-from functools import partial
 
 from cfbs.utils import (
     cfbs_dir,
@@ -44,6 +43,8 @@ from cfbs.git import (
     git_init,
     git_commit,
     git_discard_changes_in_file,
+    with_git_commit,
+    commit_after_command,
     CFBSGitError,
 )
 from cfbs.prompts import YES_NO_CHOICES, prompt_user
@@ -63,53 +64,6 @@ def _item_index(iterable, item, extra_at_end=True):
             return len(iterable)
         else:
             return -1
-
-
-
-def with_git_commit(
-    successful_returns,
-    files_to_commit,
-    commit_msg,
-    positional_args_lambdas=None,
-    failed_return=False,
-):
-    def decorator(fn):
-        def decorated_fn(*args, **kwargs):
-            ret = fn(*args, **kwargs)
-
-            config = CFBSConfig.get_instance()
-            if not config["git"]:
-                return ret
-
-            if ret in successful_returns:
-                if positional_args_lambdas:
-                    positional_args = (
-                        l_fn(args, kwargs) for l_fn in positional_args_lambdas
-                    )
-                    msg = commit_msg % tuple(positional_args)
-                else:
-                    msg = commit_msg
-
-                try:
-                    git_commit(msg, config.non_interactive, files_to_commit)
-                except CFBSGitError as e:
-                    print(str(e))
-                    try:
-                        for file_name in files_to_commit:
-                            git_discard_changes_in_file(file_name)
-                    except CFBSGitError as e:
-                        print(str(e))
-                    else:
-                        print("Failed to commit changes, discarding them...")
-                        return failed_return
-            return ret
-
-        return decorated_fn
-
-    return decorator
-
-
-commit_after_command = partial(with_git_commit, (0,), ("cfbs.json",), failed_return=0)
 
 
 def pretty_command(filenames: list, check: bool, keep_order: bool) -> int:
